+++ lib/constants/translations/app_string_constants.dart
@@ -1,0 +1,4 @@
+class AppStringConstants {
+  static const int amountOfTips = 5;
+  static const String descriptionGap = '\n';
+}