+++ lib/constants/translations/app_string_data.dart
@@ -5,6 +5,7 @@ class AppStringData {
   static final Map<String, Map<String, String>> values = {
     'en': {
       'roundText': 'Round ${AppStrings.placeholderText}',
+      'endOfRoundText': 'End Of Round ${AppStrings.placeholderText}',
       'saveGame': 'Save Game',
       'restartGame': 'Restart Game',
       'gameOver': 'Game Over',
@@ -30,6 +31,13 @@ class AppStringData {
       'alreadyPurchased': 'Purchased',
       'noItemSelected': 'No item selected',
       'potentialPurchaseCount': '${AppStrings.placeholderText}/${AppStrings.placeholderText} purchased',
+
+      // Tips
+      'gameTip0': 'Injure flying enemies by throwing others into them or by tapping',
+      'gameTip1': 'Dimensia is a hidden gem of a game',
+      'gameTip2': 'Be careful of throwing enemies into your wall to avoid damaging it',
+      'gameTip3': 'Flying enemies can not be dragged, but they can be tapped and collided with',
+      'gameTip4': 'Ensure to use your gold to upgrade your defenses between rounds',
 
       // Purchasables
       'woodenWallName': 'Wooden Wall',