+++ lib/constants/translations/app_strings.dart
@@ -1,5 +1,7 @@
+import 'package:defend_your_flame/constants/translations/app_string_constants.dart';
 import 'package:defend_your_flame/constants/translations/app_string_data.dart';
 import 'package:defend_your_flame/constants/translations/app_string_helper.dart';
+import 'package:defend_your_flame/helpers/global_vars.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
@@ -53,6 +55,7 @@ class AppStrings {
 
   // Below this are declarations of all the strings used in the app, above is setup.
   String roundText(int round) => AppStringHelper.insertNumber(getValue('roundText'), round);
+  String endOfRoundText(int round) => AppStringHelper.insertNumber(getValue('endOfRoundText'), round);
   String get restartGame => getValue('restartGame');
   String get gameOver => getValue('gameOver');
   String get gameOverRoundText => getValue('gameOverRoundText');
@@ -92,4 +95,6 @@ class AppStrings {
   String get stoneWallName => getValue('stoneWallName');
   String get stoneWallDescription => getValue('stoneWallDescription');
   String get stoneWallQuote => getValue('stoneWallQuote');
+
+  String getRandomTip() => getValue('gameTip${GlobalVars.rand.nextInt(AppStringConstants.amountOfTips)}');
 }